Team Profile Page has been redirected into teamHome. Changed redirection in teamJoin to teamHome. Got rid of teamProfile. Removed mentions of teamProfile in the Dashboard Confessional
